requirements.txt
# Core Framework & Server
fastapi==0.115.12
uvicorn[standard]==0.33.0
pydantic~=2.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
typing-extensions~=4.13.0

# Database & ORM
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
Mako~=1.3.0
greenlet~=3.1.0

# Authentication
fastapi-users[sqlalchemy]==13.0.0
email-validator==2.1.2
passlib[bcrypt]~=1.7.4

# Caching
redis[hiredis]==5.2.1
